Split meta_adaptation.py into blackjax/adaptation/meta/ subpackage
Finding: The 2801-line monolith mixed calibration constants, NamedTuples, single-chain
signals, multi-chain detection, router functions, schedule, and two large builder
functions into one file, making the decision logic hard to navigate. The `_compute_s_gap`
function was also referenced in builders.py but missing from the initial import list —
caught by mypy. Two test-only constants (`_W_BRANCH_NULL_EDGE_TW_FACTOR`,
`_W_BRANCH_PSI_FLOOR`) were not included in the shim's re-export list.

Fix: Pure move into 9-file subpackage with clean dependency graph (_calibration has
no blackjax deps; _state ← metrics; _signals/_detection/_router/_schedule ← _calibration;
builders ← all; verdict ← _state + _calibration). meta_adaptation.py is now a
back-compat shim that fires DeprecationWarning (stacklevel=1) and re-exports the
full public surface. staged_adaptation.py imports updated to the new canonical paths.
pytest.ini gets a filterwarning to suppress the shim DeprecationWarning during the
existing test run (the test split in commit 2 will remove the shim import from tests).
100 meta tests pass; pre-commit clean; import parity verified.
