+++ pytest.ini
@@ -2,3 +2,4 @@
 testpaths= "tests"
 filterwarnings =
     error
+    ignore::DeprecationWarning:blackjax.adaptation.meta_adaptation
